[REF] mrp_bom_line_net_qty, mrp_bom_weight [IMP] divide by zero
