+++ mrp_bom_line_net_qty/i18n/fr.po
@@ -6,8 +6,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 12.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-11-03 14:06+0000\n"
-"PO-Revision-Date: 2023-11-03 14:06+0000\n"
+"POT-Creation-Date: 2024-02-26 14:48+0000\n"
+"PO-Revision-Date: 2024-02-26 14:48+0000\n"
 "Last-Translator: <>\n"
 "Language-Team: \n"
 "MIME-Version: 1.0\n"
@@ -16,19 +16,9 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. module: mrp_bom_line_net_qty
-#: model:ir.model,name:mrp_bom_line_net_qty.model_mrp_bom
-msgid "Bill of Material"
-msgstr "Fiche technique"
-
-#. module: mrp_bom_line_net_qty
 #: model:ir.model,name:mrp_bom_line_net_qty.model_mrp_bom_line
 msgid "Bill of Material Line"
 msgstr "Ligne de fiche technique"
-
-#. module: mrp_bom_line_net_qty
-#: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom__diff_bom_qty_and_net_quantities
-msgid "Diff Bom Qty And Net Quantities"
-msgstr ""
 
 #. module: mrp_bom_line_net_qty
 #: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom_line__diff_product_qty_gross_net
@@ -36,34 +26,14 @@ msgid "Diff Product Qty Gross Net"
 msgstr ""
 
 #. module: mrp_bom_line_net_qty
-#: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom__display_set_quantity_with_net_quantities
-msgid "Display Set Quantity With Net Quantities"
-msgstr ""
-
-#. module: mrp_bom_line_net_qty
-#: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom_line__line_gross_weight
-msgid "Gross weight"
-msgstr "Poids brut"
-
-#. module: mrp_bom_line_net_qty
-#: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom__bom_components_total_gross_weight
-msgid "Gross weight of components"
-msgstr "Poids brut total"
-
-#. module: mrp_bom_line_net_qty
-#: model_terms:ir.ui.view,arch_db:mrp_bom_line_net_qty.view_mrp_bom_form
-msgid "It will set BoM quantity with sum of net weight quantities"
-msgstr "Cela définira la quantité de la recette avec la somme des poids nets des matières premières"
-
-#. module: mrp_bom_line_net_qty
 #: model_terms:ir.ui.view,arch_db:mrp_bom_line_net_qty.view_mrp_bom_form
 msgid "It will set product gross quantity with net quantity and loss percentage"
-msgstr ""
+msgstr "Cela définira la quantité brut de produits grâce à la quantité nette et le pourcentage de perte."
 
 #. module: mrp_bom_line_net_qty
 #: model_terms:ir.ui.view,arch_db:mrp_bom_line_net_qty.view_mrp_bom_form
 msgid "It will set product net quantity with gross quantity and loss percentage"
-msgstr "Cela définira la quantité de produits grâce à la quantité brute et le pourcentage de perte."
+msgstr "Cela définira la quantité nette de produits grâce à la quantité brute et le pourcentage de perte."
 
 #. module: mrp_bom_line_net_qty
 #: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom_line__loss_percentage
@@ -73,22 +43,7 @@ msgstr "Perte %"
 #. module: mrp_bom_line_net_qty
 #: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom_line__product_qty_net
 msgid "Net quantity"
-msgstr "Qté net"
-
-#. module: mrp_bom_line_net_qty
-#: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom_line__line_net_weight
-msgid "Net weight"
-msgstr "Poids net"
-
-#. module: mrp_bom_line_net_qty
-#: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom_line__line_net_weight_percentage
-msgid "Net weight %"
-msgstr "Poids net %"
-
-#. module: mrp_bom_line_net_qty
-#: model:ir.model.fields,field_description:mrp_bom_line_net_qty.field_mrp_bom__bom_components_total_net_weight
-msgid "Net weight of components"
-msgstr "Poids net total"
+msgstr "Qté nette"
 
 #. module: mrp_bom_line_net_qty
 #: model:ir.model.fields,help:mrp_bom_line_net_qty.field_mrp_bom_line__loss_percentage
@@ -96,11 +51,14 @@ msgid "Percentage loss, for example, when grating carrots"
 msgstr "Pourcentage de perte, après avoir râper des carottes par exemple."
 
 #. module: mrp_bom_line_net_qty
+#: code:addons/mrp_bom_line_net_qty/models/mrp_bom_line.py:58
+#, python-format
+msgid "Setting gross quantity with 100% loss makes no sense.\n"
+" Change this value."
+msgstr "Définir une quantité brute avec 100% de perte n'a pas de sens.\n"
+" Changer cette valeur de pourcentage."
+
+#. module: mrp_bom_line_net_qty
 #: model:ir.model.fields,help:mrp_bom_line_net_qty.field_mrp_bom_line__product_qty_net
 msgid "Weight after any preparation, e.g. weight of carrots after they are grated "
 msgstr "Poids net après une préparation, après avoir râper des carottes par exemple."
-
-#. module: mrp_bom_line_net_qty
-#: model_terms:ir.ui.view,arch_db:mrp_bom_line_net_qty.view_mrp_bom_form
-msgid "⇑ Update BoM quantity"
-msgstr "⇑ Mettre à jour la quantité de la recette"
