+++ mrp_bom_weight/README.rst
@@ -1,0 +1,67 @@
+===================================
+MRP BoM Line Net and Gross Quantity
+===================================
+
+.. 
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! source digest: sha256:7bdb53898fa750ea53d7444d99188cb32ae031914c658cbc20a7ed0ef190064c
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-grap%2Fgrap--odoo--custom-lightgray.png?logo=github
+    :target: https://github.com/grap/grap-odoo-custom/tree/12.0/mrp_bom_line_net_qty
+    :alt: grap/grap-odoo-custom
+
+|badge1| |badge2| |badge3|
+
+Handle gross and net quantity on BoM lines thanks to loss percentage
+Fill quantity (original one is 'gross quantity') : it set automatically 'net quantity'
+Fill loss percentage, Odoo will show two buttons to calculate net or 'gross quantity'
+You can also set 'net quantity' and calculate 'gross quantity' with loss percentage
+
+.. figure:: https://raw.githubusercontent.com/grap/grap-odoo-custom/12.0/mrp_bom_line_net_qty/static/description/mrp_bom_line_net_qty.gif
+
+/!\ Line with difference with theorical values are set in yellow
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/grap/grap-odoo-custom/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us to smash it by providing a detailed and welcomed
+`feedback <https://github.com/grap/grap-odoo-custom/issues/new?body=module:%20mrp_bom_line_net_qty%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* GRAP
+
+Contributors
+~~~~~~~~~~~~
+
+* Quentin Dupont (quentin.dupont@grap.coop)
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `grap/grap-odoo-custom <https://github.com/grap/grap-odoo-custom/tree/12.0/mrp_bom_line_net_qty>`_ project on GitHub.
+
+You are welcome to contribute.
