+++ mrp_bom_weight/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Quentin Dupont (quentin.dupont@grap.coop)
