+++ mrp_bom_weight/readme/ROADMAP.rst
@@ -1,0 +1,1 @@
+How to handle weight units when it's not kg for a component or BoM quantity ?
