+++ mrp_bom_weight/views/view_mrp_bom.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (C) 2023 - Today: GRAP (http://www.grap.coop)
+@author: Quentin DUPONT (quentin.dupont@grap.coop)
+License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+-->
+<odoo>
+
+    <!-- ************************************************************ -->
+    <!-- MRP Bill of Materials - Form -->
+    <!-- ************************************************************ -->
+
+    <record id="view_mrp_bom_form" model="ir.ui.view">
+        <field name="model">mrp.bom</field>
+        <field name="inherit_id" ref="mrp.mrp_bom_form_view" />
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='product_uom_id']/.." position="after">
+              <field name="diff_bom_qty_and_net_quantities" invisible="1"/>
+              <field name="display_set_quantity_with_net_quantities" invisible="1"/>
+              <label for="bom_components_total_net_weight"/>
+              <div>
+                <field name="bom_components_total_net_weight" class="oe_inline"/>
+                <button name="set_bom_quantity_with_net_quantities" string="⇑ Update BoM quantity"
+                  type="object" class="oe_link"
+                  help="It will set BoM quantity with sum of net weight quantities"
+                  attrs="{'invisible': ['|', ('diff_bom_qty_and_net_quantities', '=', 0), ('display_set_quantity_with_net_quantities', '=', False)]}"/>
+              </div>
+              <field name="bom_components_total_gross_weight" class="oe_inline"/>
+            </xpath>
+            <!-- BoM Lines -->
+            <xpath expr="//field[@name='bom_line_ids']/tree/field[@name='product_uom_id']" position="after">
+              <field name="line_gross_weight" groups="base.group_no_one"/>
+              <field name="line_net_weight" sum="Net weight"/>
+              <field name="line_net_weight_percentage" widget="percentage"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
