+++ setup/mrp_bom_weight/odoo/addons/mrp_bom_weight
@@ -1,0 +1,1 @@
+../../../../mrp_bom_weight
